fix: implemented dor's changes
- remove tarball files
- split 'services' into middleware, api-service and transformation-service
- use reusable const variables instead of string literals
- remove chalk
- use SDK logger for all logging
