+++ apps/quickstart-integrations-monday-code/package.json
@@ -12,7 +12,6 @@
   "license": "ISC",
   "dependencies": {
     "@mondaycom/apps-sdk": "^0.1.3",
-    "chalk": "^5.3.0",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.1",
