+++ apps/quickstart-integrations-monday-code/src/helpers.js
@@ -1,14 +1,16 @@
 import { EnvironmentVariablesManager } from "@mondaycom/apps-sdk";
+
 const secretManager = new EnvironmentVariablesManager();
 const DEVELOPMENT_ENV = "development";
+const NODE_ENV = "NODE_ENV"
 
 export const isDevelopmentEnv = () => {
-  const currentEnv = (getSecret("NODE_ENV") || DEVELOPMENT_ENV).toLowerCase();
+  const currentEnv = (getSecret(NODE_ENV) || DEVELOPMENT_ENV).toLowerCase();
   return currentEnv === "development";
 };
 
 export const getEnv = () =>
-  (getSecret("NODE_ENV") || DEVELOPMENT_ENV).toLowerCase();
+  (getSecret(NODE_ENV) || DEVELOPMENT_ENV).toLowerCase();
 
 export const getSecret = (secretKey, options = {}) => {
   const secret = secretManager.get(secretKey, options);
