--- apps/quickstart-integrations-monday-code/src/services.js
@@ -1,79 +1,0 @@
-import initMondayClient from "monday-sdk-js";
-import jwt from "jsonwebtoken";
-import { getSecret } from "./helpers.js";
-
-export const authorizeRequest = (req, res, next) => {
-  try {
-    let { authorization } = req.headers;
-    if (!authorization && req.query) {
-      authorization = req.query.token;
-    }
-    const signingSecret = getSecret("MONDAY_SIGNING_SECRET");
-    const { accountId, userId, backToUrl, shortLivedToken } = jwt.verify(
-      authorization,
-      signingSecret
-    );
-    req.session = { accountId, userId, backToUrl, shortLivedToken };
-    next();
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: "not authenticated" });
-  }
-};
-
-export const getColumnValue = async (token, itemId, columnId) => {
-  try {
-    const mondayClient = initMondayClient();
-    mondayClient.setToken(token);
-
-    const query = `query($itemId: [Int], $columnId: [String]) {
-        items (ids: $itemId) {
-          column_values(ids:$columnId) {
-            value
-          }
-        }
-      }`;
-    const variables = { columnId, itemId };
-
-    const response = await mondayClient.api(query, { variables });
-    return response.data.items[0].column_values[0].value;
-  } catch (err) {
-    console.error(err);
-  }
-};
-
-export const changeColumnValue = async (
-  token,
-  boardId,
-  itemId,
-  columnId,
-  value
-) => {
-  try {
-    const mondayClient = initMondayClient({ token });
-
-    const query = `mutation change_column_value($boardId: Int!, $itemId: Int!, $columnId: String!, $value: JSON!) {
-        change_column_value(board_id: $boardId, item_id: $itemId, column_id: $columnId, value: $value) {
-          id
-        }
-      }
-      `;
-    const variables = { boardId, columnId, itemId, value };
-
-    const response = await mondayClient.api(query, { variables });
-    return response;
-  } catch (err) {
-    console.error(err);
-  }
-};
-
-export const transformText = (value, type) => {
-  switch (type) {
-    case "TO_UPPER_CASE":
-      return value.toUpperCase();
-    case "TO_LOWER_CASE":
-      return value.toLowerCase();
-    default:
-      return value.toUpperCase();
-  }
-};
