+++ apps/quickstart-integrations-monday-code/src/transformation-service.js
@@ -1,0 +1,10 @@
+export const transformText = (value, type) => {
+  switch (type) {
+    case "TO_UPPER_CASE":
+      return value.toUpperCase();
+    case "TO_LOWER_CASE":
+      return value.toLowerCase();
+    default:
+      return value.toUpperCase();
+  }
+};
